Move files - record serialisation into StateStoreFileUtils
